[Design] Added styles from Cat-API
