+++ src/components/pagination/Pagination.js
@@ -1,14 +1,15 @@
 import React from 'react'
+import './styles/pagination.css'
 
 const Pagination = ({ controlPagination }) => {
     const {pagination, changePagination} = controlPagination
 
     return (
-        <footer>
+        <footer className="justify-center mt-5">
             {
                 pagination.map((element, index) => {
                     return (
-                        <button key={index} onClick={() => changePagination(element)}>
+                        <button key={index} className="buttons-pagination" onClick={() => changePagination(element)}>
                             {element}
                         </button>
                     )
