+++ src/components/pagination/styles/pagination.css
@@ -1,0 +1,20 @@
+.justify-center{
+    display: flex;
+    justify-content: center;    
+}
+
+.buttons-pagination{
+    background: rgb(250, 250, 250);
+    border: 2px solid rgb(37, 190, 218);
+    border-radius: 5px;
+    box-shadow: none;
+    font-size: medium;
+    height: 2em;
+    margin: 0.3rem;
+    width: 4em;
+}
+
+.buttons-pagination:hover{
+    background: rgb(238, 237, 237);
+    border: 2px solid rgb(28, 163, 187);
+}
