+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import CatAppRouter from './routers/CatAppRouter';
-import './index.css';
+import './styles/index.css';
 
 ReactDOM.render(
   <CatAppRouter />,
